+++ frontend-javafx/settings.gradle
