+++ frontend-javafx/src/main/java/io/github/aosn/camp2016/ui/logic/EntryServiceImpl.java
