+++ frontend-javafx/src/main/java/io/github/aosn/camp2016/ui/service/HttpClient.java
