+++ frontend-javafx/src/main/java/io/github/aosn/camp2016/ui/stub/StubData.java
