+++ frontend-javafx/src/main/java/io/github/aosn/util/tryable/Failure.java
