+++ frontend-javafx/src/main/java/io/github/aosn/util/tryable/Success.java
