+++ frontend-javafx/src/main/resources/styles/main.css
